+++ packages/cli/src/types.ts
@@ -1,7 +1,6 @@
 import type { BlockManifest } from '@appsemble/types';
 import type { PackageJson } from 'read-pkg-up';
 import type { JsonObject } from 'type-fest';
-import type { URL as URL_, URLSearchParams as URLSearchParams_ } from 'url';
 
 /**
  * THe base arguments from the command line.
@@ -72,12 +71,4 @@ export interface BlockConfig
 
 export interface MonoRepoPackageJson extends PackageJson {
   appsembleServer?: JsonObject;
-}
-
-declare global {
-  // https://github.com/DefinitelyTyped/DefinitelyTyped/issues/34960
-  // eslint-disable-next-line no-redeclare
-  const URL: typeof URL_;
-  // eslint-disable-next-line no-redeclare
-  const URLSearchParams: typeof URLSearchParams_;
 }