+++ packages/server/src/models/Member.ts
@@ -1,7 +1,11 @@
-import { roles } from '@appsemble/utils';
+import { Role, roles } from '@appsemble/utils';
 import { DataTypes, Model, Sequelize } from 'sequelize';
 
 export default class Member extends Model {
+  role: Role;
+
+  OrganizationId: string;
+
   static initialize(sequelize: Sequelize): void {
     Member.init(
       {