+++ CHANGELOG.md
@@ -7,6 +7,20 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.12.8] - 2020-04-29
+
+### Fixed
+
+- Server: Fix connection to database with self-signed certificate.
+
+## [0.12.7] - 2020-04-29
+
+### Fixed
+
+- CI: Fix the release process for production.
+
+## [0.12.6] - 2020-04-29
+
 ### Added
 
 - App: Add meta description tag.
@@ -16,7 +30,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Studio: Add asset management page for apps. This can be used to preview, download, delete, and
   update assets.
 
-### Fixes
+### Fixed
 
 - App: Fix app theme color.
 - Studio: Add asset management page for apps. This can be used to preview, download, delete, and